Moves vm.NewContext to context.New
With this commit, we break the circular dependency between contracts &
vm by moving out the creation of the Context object. This object needs
to define a function to get registered address and a function to
compute the tobin tax.

Alternatively (as in a previous commit) we can get that separation by
injection those function in the ContractsContext. But this approach is
simpler, and also completely removes the notion of Tobin tax from the vm
package.
