+++ core/state_prefetcher.go
@@ -21,10 +21,10 @@ import (
 
 	"github.com/celo-org/celo-blockchain/common"
 	"github.com/celo-org/celo-blockchain/consensus"
-	"github.com/celo-org/celo-blockchain/contracts"
 	"github.com/celo-org/celo-blockchain/core/state"
 	"github.com/celo-org/celo-blockchain/core/types"
 	"github.com/celo-org/celo-blockchain/core/vm"
+	"github.com/celo-org/celo-blockchain/core/vm/context"
 	"github.com/celo-org/celo-blockchain/params"
 )
 
@@ -87,8 +87,8 @@ func precacheTransaction(config *params.ChainConfig, bc vm.ChainContext, author 
 		return err
 	}
 	// Create the EVM and execute the transaction
-	context := vm.NewEVMContext(msg, header, bc, contracts.Context, author)
-	vm := vm.NewEVM(context, statedb, config, cfg)
+	ctx := context.New(msg, header, bc, author)
+	vm := vm.NewEVM(ctx, statedb, config, cfg)
 
 	_, err = ApplyMessage(vm, msg, gaspool)
 	return err
